@@ -8,21 +8,12 @@ namespace Query
     {
         public static int GetCount(int userId)
         {
-            return Sql.ExecuteScalar<int>("Trash_GetCount",
-                new Dictionary<string, object>()
-                {
-                    {"userId", userId }
-                }
-            );
+            return Sql.ExecuteScalar<int>("Trash_GetCount", new { userId });
         }
 
         public static Tuple<List<Models.Book>, List<Models.Chapter>, List<Models.Entry>> GetList(int userId)
         {
-            using (var sql = new Connection("Trash_GetList",
-                new Dictionary<string, object>()
-                {
-                    {"userId", userId }
-                }))
+            using (var sql = new Connection("Trash_GetList", new { userId }))
             {
                 var reader = sql.PopulateMultiple();
                 var books = reader.Read<Models.Book>().AsList();
@@ -34,21 +25,11 @@ namespace Query
         }
         public static void Empty(int userId)
         {
-            Sql.ExecuteNonQuery("Trash_Empty",
-                new Dictionary<string, object>()
-                {
-                    {"userId", userId }
-                }
-            );
+            Sql.ExecuteNonQuery("Trash_Empty", new { userId });
         }
         public static void RestoreAll(int userId)
         {
-            Sql.ExecuteNonQuery("Trash_RestoreAll",
-                new Dictionary<string, object>()
-                {
-                    {"userId", userId }
-                }
-            );
+            Sql.ExecuteNonQuery("Trash_RestoreAll", new { userId });
         }
     }
 }
